+++ src/global/SilentLogin/index.tsx
@@ -1,0 +1,11 @@
+import SilentLoginAfterLogin from '@/global/SilentLogin/SilentLoginAfterLogin';
+import SilentLoginAfterRevisit from '@/global/SilentLogin/SilentLoginAfterRevisit';
+
+const SilentLogin = () => (
+  <>
+    <SilentLoginAfterLogin />
+    <SilentLoginAfterRevisit />
+  </>
+);
+
+export default SilentLogin;
